Use nightly-compatible branch of rkyv
Upstream rkyv (0.7) supports the hashbrown crate explicitly (0.12).
In turn, the hashbrown crate requires a version of ahash that is not
compatible with latest rust nightlies.

There is an update to ahash; and there is an update to hashbrown; but
rkyv doesn't want to pull those in because supporting hashbrown @ 0.12
is part of their public API.

So, we point to a branch where rkyv supports the latest and greatest of
hashbrown and ahash; it is the same as 0.7 in all other respects. Once
rkyv releases 0.8 (which is a few months away), then we can switch back
to regular releases and everything works again.

Gotta love it.